rename: enum 클래스 패키지 이름 수정
