--- src/main/java/racingcar/util/RandomNumber.java
@@ -1,5 +1,0 @@
-package racingcar.util;
-
-public class RandomNumber {
-    //테스트코드하기 좋게 수정
-}
