+++ src/main/java/racingcar/validator/NameDuplicateValidator.java
@@ -1,6 +1,6 @@
 package racingcar.validator;
 
-import racingcar.util.ExceptionMessage;
+import racingcar.constant.ExceptionMessage;
 
 import java.util.List;
